[actor #173] rename immutable::Signature to SignatureRef!
