+++ git-actor/src/lib.rs
@@ -1,10 +1,12 @@
-//! This crate provides ways of identifying an actor within the git repository both in shared/immutable and mutable variants.
+//! This crate provides ways of identifying an actor within the git repository both in shared/signature_ref and mutable variants.
 #![forbid(unsafe_code)]
 #![deny(rust_2018_idioms, missing_docs)]
-use bstr::BString;
+use bstr::{BStr, BString};
 
-pub mod immutable;
-mod signature;
+///
+pub mod signature;
+///
+pub mod signature_ref;
 
 const SPACE: &[u8; 1] = b" ";
 
@@ -19,6 +21,21 @@ pub struct Signature {
     /// The actor's email.
     pub email: BString,
     /// The time stamp at which the signature is performed.
+    pub time: Time,
+}
+
+/// A immutable signature is created by an actor at a certain time.
+///
+/// Note that this is not a cryptographical signature.
+#[derive(PartialEq, Eq, Debug, Hash, Ord, PartialOrd, Clone)]
+#[cfg_attr(feature = "serde1", derive(serde::Serialize, serde::Deserialize))]
+pub struct SignatureRef<'a> {
+    /// The actor's name.
+    #[cfg_attr(feature = "serde1", serde(borrow))]
+    pub name: &'a BStr,
+    /// The actor's email.
+    pub email: &'a BStr,
+    /// The time stamp at which the signature was performed.
     pub time: Time,
 }
