+++ git-actor/src/types.rs
@@ -1,21 +1,6 @@
 use std::io;
 
-use crate::{Sign, Signature, Time, SPACE};
-
-impl Signature {
-    /// An empty signature, similar to 'null'.
-    pub fn empty() -> Self {
-        Signature {
-            name: Default::default(),
-            email: Default::default(),
-            time: Time {
-                time: 0,
-                offset: 0,
-                sign: Sign::Plus,
-            },
-        }
-    }
-}
+use crate::{Sign, Time, SPACE};
 
 impl Time {
     /// Serialize this instance to `out` in a format suitable for use in header fields of serialized git commits or tags.
