--- git-actor/tests/mutable/mod.rs
@@ -1,1 +1,0 @@
-mod signature;
