+++ git-object/src/lib.rs
@@ -1,4 +1,4 @@
-//! This crate provides types for [read-only git objects][immutable::Object] backed by bytes provided in git's serialization format
+//! This crate provides types for [read-only git objects][signature_ref::Object] backed by bytes provided in git's serialization format
 //! as well as [mutable versions][mutable::Object] of these. The latter can be serialized into git's serialization format for objects.
 #![forbid(unsafe_code)]
 #![deny(rust_2018_idioms, missing_docs)]
