+++ git-object/src/types.rs
@@ -56,7 +56,7 @@ impl fmt::Display for Kind {
 pub mod tree {
     /// The mode of items storable in a tree, similar to the file mode on a unix file system.
     ///
-    /// Used in [mutable::Entry][crate::mutable::tree::Entry] and [immutable::Entry][crate::immutable::tree::Entry].
+    /// Used in [mutable::Entry][crate::mutable::tree::Entry] and [signature_ref::Entry][crate::signature_ref::tree::Entry].
     #[derive(Clone, Copy, PartialEq, Eq, Debug, Ord, PartialOrd, Hash)]
     #[repr(u16)]
     #[cfg_attr(feature = "serde1", derive(serde::Serialize, serde::Deserialize))]
