+++ git-ref/src/store/file/log/mod.rs
@@ -13,7 +13,7 @@ pub struct Line<'a> {
     pub new_oid: &'a BStr,
     /// The signature of the currently configured committer.
     #[cfg_attr(feature = "serde1", serde(borrow))]
-    pub signature: git_actor::immutable::Signature<'a>,
+    pub signature: git_actor::SignatureRef<'a>,
     /// The message providing details about the operation performed in this log line.
     pub message: &'a BStr,
 }
